Extract project marker traversal partial
